Move settings callbacks to sse module
