+++ pysmlight/__init__.py
@@ -5,4 +5,5 @@ __all__ = [
     "SettingsEvent",
 ]
 
-from pysmlight.web import Api2, Info, Sensors, SettingsEvent
+from pysmlight.models import SettingsEvent
+from pysmlight.web import Api2, Info, Sensors
