+++ pysmlight/sse.py
@@ -1,10 +1,12 @@
+import json
 import logging
 from typing import Callable
 
 import aiohttp
 from aiohttp_sse_client2.client import EventSource, MessageEvent
 
-from .const import Events
+from .const import Events, Pages, Settings
+from .models import SettingsEvent
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -18,11 +20,14 @@ class sseClient:
 
     def __init__(self, host: str, session: aiohttp.ClientSession):
         self.callbacks: dict[str, Callable] = {}
+        self.settings_cb: dict[Settings, Callable] = {}
         self.session = session
         self.url = f"http://{host}/events"
         self.timeout = aiohttp.ClientTimeout(
             total=None, connect=None, sock_connect=None, sock_read=30
         )
+
+        self.register_callback(Events.SAVE_PARAMS, self._handle_settings)
 
     async def client(self):
         while True:
@@ -63,3 +68,29 @@ class sseClient:
         else:
             cb = self.callbacks.pop(event.name, None)
         return cb
+
+    def _handle_settings(self, event: Events) -> None:
+        data = json.loads(event.data)
+        page = Pages(data["page"])
+        changes = data.pop("changes", None)
+        for setting in changes:
+            base = data.copy()
+            match_cb = next(
+                (
+                    cb
+                    for k, cb in self.settings_cb.items()
+                    if (page, setting) == k.value
+                ),
+                None,
+            )
+
+            if match_cb:
+                base["setting"] = {setting: changes[setting]}
+                result = SettingsEvent.from_dict(base)
+                match_cb(result)
+
+    def register_settings_cb(self, setting: Settings, cb: Callable) -> None:
+        self.settings_cb[setting] = cb
+
+    def deregister_settings_cb(self, setting: Settings) -> None:
+        self.settings_cb.pop(setting)
